+++ src/components/Header/Header.module.css
@@ -1,12 +1,3 @@
-/* html,
-body {
-  font-size: 62.5%;
-} */
-
-/* * {
-  box-sizing: border-box;
-} */
-
 .nav {
   padding: 20px 200px;
   box-sizing: border-box;
